requirements.txt
numpy
scipy
matplotlib
tqdm
Pillow
opencv-python
scikit-learn

torch
torchvision

transformers
pytorch-lightning
torchmetrics
wandb
google-genai